Delete MathUtils as it was merged with Utility class and Reorganize/Restructure
