+++ android/src/main/java/com/terrylinla/rnsketchcanvas/utils/gestureDetectors/MoveGestureDetector.java
@@ -1,4 +1,4 @@
-package com.terrylinla.utils.gestureDetectors;
+package com.terrylinla.rnsketchcanvas.utils.gestureDetectors;
 
 import android.content.Context;
 import android.graphics.PointF;
